Refactor pool to remove interfaces and update schema to support vlocity
