+++ packages/sfpowerscripts-cli/src/impl/prepare/PrepareOrgJob.ts
@@ -8,7 +8,8 @@ import DeployImpl, {
 import SFPLogger, {
   FileLogger,
   LoggerLevel,
-  Logger
+  Logger,
+  COLOR_KEY_MESSAGE
 } from "@dxatscale/sfpowerscripts.core/lib/logger/SFPLogger";
 import { Stage } from "../Stage";
 import SFPStatsSender from "@dxatscale/sfpowerscripts.core/lib/stats/SFPStatsSender";
@@ -25,6 +26,8 @@ import { Result, ok, err } from "neverthrow";
 import { ArtifactFilePaths } from "@dxatscale/sfpowerscripts.core/lib/artifacts/ArtifactFilePathFetcher";
 import RelaxIPRange from "@dxatscale/sfpowerscripts.core/lib/iprange/RelaxIPRange"
 import SourceTrackingResourceController from "../pool/SourceTrackingResourceController";
+import VlocityPackUpdateSettings from "@dxatscale/sfpowerscripts.core/lib/vlocitywrapper/VlocityPackUpdateSettings";
+import VlocityInitialInstall from "@dxatscale/sfpowerscripts.core/lib/vlocitywrapper/VlocityInitialInstall";
 
 const SFPOWERSCRIPTS_ARTIFACT_PACKAGE = "04t1P000000ka9mQAA";
 export default class PrepareOrgJob extends PoolJobExecutor {
@@ -40,7 +43,8 @@ export default class PrepareOrgJob extends PoolJobExecutor {
   async executeJob(
     scratchOrg: ScratchOrg,
     hubOrg: Org,
-    logToFilePath: string
+    logToFilePath: string,
+    logLevel: LoggerLevel
   ): Promise<Result<ScriptExecutionResult, JobError>> {
     //Install sfpowerscripts Artifact
 
@@ -104,6 +108,12 @@ export default class PrepareOrgJob extends PoolJobExecutor {
       SFPLogger.log(
         `Successfully completed Installing Package Dependencies of this repo in ${scratchOrg.alias}`
       );
+
+
+
+      //Hook Velocity Deployment
+      if(this.pool.enableVlocity)
+        await this.prepareVlocityDataPacks(scratchOrg,packageLogger,logLevel);
 
       if (this.artifacts) {
         let deploymentResult: DeploymentResult;
@@ -283,4 +293,21 @@ export default class PrepareOrgJob extends PoolJobExecutor {
       );
     }
   }
+
+  
+  //Prepare for vlocity
+  private async prepareVlocityDataPacks(scratchOrg:ScratchOrg,logger:Logger,logLevel:LoggerLevel)
+  {
+  
+    SFPLogger.log(COLOR_KEY_MESSAGE("Installing Vlocity Configurations.."),LoggerLevel.INFO,logger);
+    let vlocityPackSettingsUpdate:VlocityPackUpdateSettings = new VlocityPackUpdateSettings(null,scratchOrg.username,logger,logLevel);
+    await vlocityPackSettingsUpdate.exec(false);
+
+    let vlocityInitialInstall:VlocityInitialInstall = new VlocityInitialInstall(null,scratchOrg.username,logger,logLevel);
+    await vlocityInitialInstall.exec(false);
+    SFPLogger.log(COLOR_KEY_MESSAGE("Succesfully completed all vlocity config installation"),LoggerLevel.INFO,logger);
+  }
+
+  
+
 }
